backend/requirements.txt
aiosqlite==0.19.0
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
arrow==1.3.0
bcrypt==4.0.1
binaryornot==0.4.4
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cookiecutter==1.7.3
cryptography==41.0.4
databases==0.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
Mako==1.2.4
manage-fastapi==1.1.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
poyo==0.5.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
pytest==7.4.2
pytest-env==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0.1
questionary==1.10.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
text-unidecode==1.3
tomli==2.0.1
typer==0.4.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
